refactor(overlay): extract ShellHost bridge methods from osd.cpp (phase 5.3)
The 6 daemon-to-ShellHost bridge methods (ensurePassiveShellFor /
destroyPassiveShell / wirePassiveShellSlots / unwirePassiveShellSlots /
warmUpNotifications / ensureOsdScreenAddedConnected) accumulated in
osd.cpp during the Phase 2-4 ShellHost lift, plus the surface-sync
helpers (syncPassiveShellSurfaceState / syncPassiveShellSurfaceStateForSurface).
They're not OSD-specific — they're the bridge between the daemon and
the lib's ShellHost.

Move all 8 to overlayservice/shellhost_bridge.cpp. osd.cpp shrinks
from 889 to 693 lines (under the project's <800-line cap); the new
shellhost_bridge.cpp is 249 lines and groups the bridge code with
itself.

OSD-specific methods (onOsdDismissRequested, onOsdSlotHideCompleted,
showLayoutOsd, prepareLayoutOsdWindow, pushLayoutOsdContent,
showDisabledOsd, showNavigationOsd) stay in osd.cpp where they belong.

Build clean. 178/178 tests pass.
